+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Event handler register/unregister.
+- Read misreporting a closed connection as a timeout.
 
 ## [1.2.2] - 2026-08-11
 ### Fixed